backend/requirements.txt
# Backend Requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
onnxruntime==1.16.3
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80

# Utilities
httpx==0.26.0
python-dotenv==1.0.0

# Google Gemini AI (new SDK)
google-genai>=1.0.0
